refactor(web): move feature-flag stores to stores/, hooks to hooks/, rename -utils files
- Move assistant-feature-flag-store.ts and client-feature-flag-store.ts
  from lib/feature-flags/ to stores/ (Zustand stores belong in stores/)
- Move use-assistant-feature-flag-sync.ts, use-client-feature-flag-sync.ts,
  use-feature-flag-bus-sync.ts from lib/feature-flags/ to hooks/
  (cross-domain hooks belong in hooks/)
- Colocate use-feature-flag-bus-sync.test.tsx with its hook in hooks/
- Keep feature-flag-catalog.ts and feature-flag-registry.json in
  lib/feature-flags/ (config/infrastructure data)
- Rename redundant -utils suffix files inside utils/ directories:
  chat-utils.ts -> chat.ts, risk-utils.ts -> risk.ts,
  sub-group-utils.ts -> sub-group.ts, notification-utils.ts -> notification.ts
- Update all 50+ import paths across the codebase

Closes LUM-2025
