+++ apps/web/src/domains/chat/hooks/use-interaction-actions.ts
@@ -21,7 +21,7 @@ import { useConversationStore } from "@/domains/conversations/conversation-store
 import { useTurnStore } from "@/domains/messaging/turn-store";
 
 import { clearConfirmationByRequestId } from "@/domains/chat/hooks/send-message-utils";
-import { deriveCommandText } from "@/domains/chat/utils/chat-utils";
+import { deriveCommandText } from "@/domains/chat/utils/chat";
 import type { ChatError } from "@/domains/chat/types";
 import type { AllowlistOption, ConfirmationDecision, DirectoryScopeOption, QuestionResponseEntry, ScopeOption } from "@/domains/chat/api/event-types";
 import { submitConfirmation, submitContactPrompt, submitQuestionResponse, submitSecretResponse } from "@/domains/chat/api/interactions";
