+++ apps/web/src/domains/chat/utils/background-sub-groups.ts
@@ -1,5 +1,5 @@
-import { groupConversationsByKey } from "@/domains/chat/utils/sub-group-utils";
-import type { SubGroup } from "@/domains/chat/utils/sub-group-utils";
+import { groupConversationsByKey } from "@/domains/chat/utils/sub-group";
+import type { SubGroup } from "@/domains/chat/utils/sub-group";
 import type { Conversation } from "@/types/conversation-types";
 
 export type BackgroundSubGroup = SubGroup;
