+++ apps/web/src/domains/chat/utils/chat.ts
