+++ apps/web/src/domains/chat/utils/stream-handlers/error-handlers.ts
@@ -3,7 +3,7 @@ import {
   handleConversationError,
   stopStreaming,
 } from "@/domains/chat/hooks/stream-message-updaters";
-import { ERROR_MESSAGES } from "@/domains/chat/utils/chat-utils";
+import { ERROR_MESSAGES } from "@/domains/chat/utils/chat";
 import type { StreamHandlerContext } from "@/domains/chat/utils/stream-handlers/types";
 import type { ConversationErrorEvent, StreamErrorEvent } from "@/domains/chat/api/event-types";
