+++ apps/web/src/domains/chat/utils/sub-group.ts
