+++ apps/web/src/domains/settings/components/dev-mode-version-unlock.tsx
@@ -1,7 +1,7 @@
 import { Loader2 } from "lucide-react";
 import { useCallback, useEffect, useRef, useState } from "react";
 
-import { useAssistantFeatureFlagStore } from "@/lib/feature-flags/assistant-feature-flag-store";
+import { useAssistantFeatureFlagStore } from "@/stores/assistant-feature-flag-store";
 
 const TAP_THRESHOLD = 7;
 const MESSAGE_DURATION_MS = 2000;
