+++ apps/web/src/domains/settings/pages/general-page.tsx
@@ -21,8 +21,8 @@ import {
 } from "@/domains/settings/components/assistant-status-panel";
 
 import { useAuthStore } from "@/stores/auth-store";
-import { useClientFeatureFlagStore } from "@/lib/feature-flags/client-feature-flag-store";
-import { useAssistantFeatureFlagStore } from "@/lib/feature-flags/assistant-feature-flag-store";
+import { useClientFeatureFlagStore } from "@/stores/client-feature-flag-store";
+import { useAssistantFeatureFlagStore } from "@/stores/assistant-feature-flag-store";
 import { isLocalMode } from "@/lib/local-mode";
 import {
   applyThemePreference,
