+++ apps/web/src/domains/settings/utils/notification.ts
