+++ apps/web/src/hooks/use-assistant-feature-flag-sync.ts
@@ -3,7 +3,7 @@ import { useQuery } from "@tanstack/react-query";
 
 import { client } from "@/generated/api/client.gen";
 import { assertHasResponse } from "@/lib/api-errors";
-import { useAssistantFeatureFlagStore } from "@/lib/feature-flags/assistant-feature-flag-store";
+import { useAssistantFeatureFlagStore } from "@/stores/assistant-feature-flag-store";
 import {
   ASSISTANT_FLAG_DEFAULTS,
   flagKeyToStoreKey,
